Updated Home Page
Added Nav Bar with Home, News, Contact, About, and Games as well as outlining the nav bar and bottom link in black and aligning the botttom links to well the bottom. I have also taken the liberty of changin the link font color to white so its visable against the black background.